+++ planCharge-ihm/src/main/java/fr/gouv/agriculture/dal/ct/planCharge/ihm/model/disponibilite/NbrsJoursDAbsenceBean.java
@@ -3,6 +3,7 @@ package fr.gouv.agriculture.dal.ct.planCharge.ihm.model.disponibilite;
 import fr.gouv.agriculture.dal.ct.ihm.model.BeanException;
 import fr.gouv.agriculture.dal.ct.metier.dto.AbstractDTO;
 import fr.gouv.agriculture.dal.ct.planCharge.ihm.model.referentiels.RessourceHumaineBean;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 
 import javax.validation.constraints.NotNull;
@@ -15,14 +16,14 @@ import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
-public class NbrsJoursDAbsenceBean extends AbstractDisponibilitesRessourceBean<AbstractDTO, NbrsJoursDAbsenceBean, IntegerProperty> {
+public class NbrsJoursDAbsenceBean extends AbstractDisponibilitesRessourceBean<AbstractDTO, NbrsJoursDAbsenceBean, DoubleProperty> {
 
 
     // Fields :
 
     // Constructors :
 
-    public NbrsJoursDAbsenceBean(@NotNull RessourceHumaineBean ressourceHumaineBean, @NotNull Map<LocalDate, IntegerProperty> calendrier) {
+    public NbrsJoursDAbsenceBean(@NotNull RessourceHumaineBean ressourceHumaineBean, @NotNull Map<LocalDate, DoubleProperty> calendrier) {
         super(ressourceHumaineBean, calendrier);
     }
