+++ planCharge-metier/src/main/java/fr/gouv/agriculture/dal/ct/planCharge/metier/dao/disponibilite/xml/DisponibilitesXmlWrapper.java
@@ -74,10 +74,10 @@ public class DisponibilitesXmlWrapper {
     public Disponibilites extract() throws DaoException {
 
         // Absences :
-        Map<RessourceHumaine, Map<LocalDate, Integer>> absences = new TreeMap<>(); // TreeMap au lieu de HashMap juste pour trier afin de faciliter le débogage.
+        Map<RessourceHumaine, Map<LocalDate, Double>> absences = new TreeMap<>(); // TreeMap au lieu de HashMap juste pour trier afin de faciliter le débogage.
         for (String trigrammeRsrcHum : absencesXmlWrapper.keySet()) {
             RessourceHumaine ressourceHumaine = ressourceHumaineDao.load(trigrammeRsrcHum);
-            Map<LocalDate, Integer> calendrier = absencesXmlWrapper.get(trigrammeRsrcHum).extract();
+            Map<LocalDate, Double> calendrier = absencesXmlWrapper.get(trigrammeRsrcHum).extract();
             absences.put(ressourceHumaine, calendrier);
         }
